# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Catan Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/games": {
      "get": {
        "tags": [
          "games"
        ],
        "summary": "List Games",
        "description": "List the user's games and open lobbies they could join.\n\nArgs:\n    current_user: Authenticated user (injected dependency).\n    db: Database session (injected dependency).\n\nReturns:\n    The user's games and joinable lobbies.",
        "operationId": "list_games_games_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameList"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "games"
        ],
        "summary": "Create Game",
        "description": "Open a lobby hosted by the user, optionally with bots already seated.\n\nArgs:\n    payload: Table size and bot count.\n    current_user: Authenticated user (injected dependency).\n    db: Database session (injected dependency).\n\nReturns:\n    The new lobby.",
        "operationId": "create_game_games_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/games/{game_id}": {
      "get": {
        "tags": [
          "games"
        ],
        "summary": "Get Game",
        "description": "Get a game and the user's view of the board. Clients poll this.\n\nArgs:\n    game_id: Game id.\n    current_user: Authenticated user (injected dependency).\n    db: Database session (injected dependency).\n\nReturns:\n    The game; spectators see no private hands.\n\nRaises:\n    HTTPException: 404 if the game does not exist.",
        "operationId": "get_game_games__game_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Game Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/games/{game_id}/join": {
      "post": {
        "tags": [
          "games"
        ],
        "summary": "Join Game",
        "description": "Take a free seat in a lobby.\n\nArgs:\n    game_id: Game id.\n    current_user: Authenticated user (injected dependency).\n    db: Database session (injected dependency).\n\nReturns:\n    The lobby.\n\nRaises:\n    HTTPException: 404 if missing; 409 if started or full.",
        "operationId": "join_game_games__game_id__join_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Game Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/games/{game_id}/leave": {
      "post": {
        "tags": [
          "games"
        ],
        "summary": "Leave Game",
        "description": "Leave a game: frees a lobby seat (the host closes the lobby) or hands\na seat in a running game to a bot.\n\nArgs:\n    game_id: Game id.\n    current_user: Authenticated user (injected dependency).\n    db: Database session (injected dependency).\n\nReturns:\n    Empty 204 response.\n\nRaises:\n    HTTPException: 404 if missing or not seated.",
        "operationId": "leave_game_games__game_id__leave_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Game Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/games/{game_id}/bots": {
      "post": {
        "tags": [
          "games"
        ],
        "summary": "Add Bot",
        "description": "Seat a random bot (host only).\n\nArgs:\n    game_id: Game id.\n    current_user: Authenticated user (injected dependency).\n    db: Database session (injected dependency).\n\nReturns:\n    The lobby.\n\nRaises:\n    HTTPException: 403 if not the host; 404 if missing; 409 if started or full.",
        "operationId": "add_bot_games__game_id__bots_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Game Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/games/{game_id}/seats/{seat}": {
      "delete": {
        "tags": [
          "games"
        ],
        "summary": "Remove Seat",
        "description": "Remove a bot or another player from the lobby (host only).\n\nArgs:\n    game_id: Game id.\n    seat: Seat to clear.\n    current_user: Authenticated user (injected dependency).\n    db: Database session (injected dependency).\n\nReturns:\n    The lobby.\n\nRaises:\n    HTTPException: 403 if not the host; 404 if missing or empty; 409 if started.",
        "operationId": "remove_seat_games__game_id__seats__seat__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Game Id"
            }
          },
          {
            "name": "seat",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Seat"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/games/{game_id}/start": {
      "post": {
        "tags": [
          "games"
        ],
        "summary": "Start Game",
        "description": "Start the game with whoever is seated (host only, two or more players).\n\nArgs:\n    game_id: Game id.\n    current_user: Authenticated user (injected dependency).\n    db: Database session (injected dependency).\n\nReturns:\n    The started game.\n\nRaises:\n    HTTPException: 403 if not the host; 404 if missing; 409 if started or too few players.",
        "operationId": "start_game_games__game_id__start_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Game Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/games/{game_id}/actions": {
      "post": {
        "tags": [
          "games"
        ],
        "summary": "Submit Action",
        "description": "Make a move; bots then play until a human is up.\n\nArgs:\n    game_id: Game id.\n    payload: The move.\n    current_user: Authenticated user (injected dependency).\n    db: Database session (injected dependency).\n\nReturns:\n    The game after the move and any bot moves.\n\nRaises:\n    HTTPException: 400 if the move is illegal; 403 if not seated; 404 if\n        missing; 409 if the game is not in progress.",
        "operationId": "submit_action_games__game_id__actions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "game_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Game Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GameActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BuildCity": {
        "properties": {
          "type": {
            "type": "string",
            "const": "build_city",
            "title": "Type",
            "default": "build_city"
          },
          "node": {
            "type": "integer",
            "title": "Node"
          }
        },
        "type": "object",
        "required": [
          "node"
        ],
        "title": "BuildCity",
        "description": "upgrade an owned settlement to a city."
      },
      "BuildRoad": {
        "properties": {
          "type": {
            "type": "string",
            "const": "build_road",
            "title": "Type",
            "default": "build_road"
          },
          "edge": {
            "prefixItems": [
              {
                "type": "integer"
              },
              {
                "type": "integer"
              }
            ],
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Edge"
          }
        },
        "type": "object",
        "required": [
          "edge"
        ],
        "title": "BuildRoad",
        "description": "place a road on an edge."
      },
      "BuildSettlement": {
        "properties": {
          "type": {
            "type": "string",
            "const": "build_settlement",
            "title": "Type",
            "default": "build_settlement"
          },
          "node": {
            "type": "integer",
            "title": "Node"
          }
        },
        "type": "object",
        "required": [
          "node"
        ],
        "title": "BuildSettlement",
        "description": "place a settlement on a node."
      },
      "BuyDevCard": {
        "properties": {
          "type": {
            "type": "string",
            "const": "buy_dev_card",
            "title": "Type",
            "default": "buy_dev_card"
          }
        },
        "type": "object",
        "title": "BuyDevCard",
        "description": "buy the top development card."
      },
...
      "Discard": {
        "properties": {
          "type": {
            "type": "string",
            "const": "discard",
            "title": "Type",
            "default": "discard"
          },
          "resources": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/Resource"
            },
            "type": "object",
            "title": "Resources"
          }
        },
        "type": "object",
        "required": [
          "resources"
        ],
        "title": "Discard",
        "description": "return half a hand of more than seven cards after a 7 is rolled."
      },
...
      "GameActionRequest": {
        "properties": {
          "action": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/RollDice"
              },
              {
                "$ref": "#/components/schemas/Discard"
              },
              {
                "$ref": "#/components/schemas/MoveRobber"
              },
              {
                "$ref": "#/components/schemas/BuildRoad"
              },
              {
                "$ref": "#/components/schemas/BuildSettlement"
              },
              {
                "$ref": "#/components/schemas/BuildCity"
              },
              {
                "$ref": "#/components/schemas/BuyDevCard"
              },
              {
                "$ref": "#/components/schemas/PlayKnight"
              },
              {
                "$ref": "#/components/schemas/PlayRoadBuilding"
              },
              {
                "$ref": "#/components/schemas/PlayYearOfPlenty"
              },
              {
                "$ref": "#/components/schemas/PlayMonopoly"
              },
              {
                "$ref": "#/components/schemas/MaritimeTrade"
              },
              {
                "$ref": "#/components/schemas/ProposeTrade"
              },
              {
                "$ref": "#/components/schemas/RespondTrade"
              },
              {
                "$ref": "#/components/schemas/ConfirmTrade"
              },
              {
                "$ref": "#/components/schemas/CancelTrade"
              },
              {
                "$ref": "#/components/schemas/EndTurn"
              }
            ],
            "title": "Action",
            "discriminator": {
              "propertyName": "type",
              "mapping": {
                "build_city": "#/components/schemas/BuildCity",
                "build_road": "#/components/schemas/BuildRoad",
                "build_settlement": "#/components/schemas/BuildSettlement",
                "buy_dev_card": "#/components/schemas/BuyDevCard",
                "cancel_trade": "#/components/schemas/CancelTrade",
                "confirm_trade": "#/components/schemas/ConfirmTrade",
                "discard": "#/components/schemas/Discard",
                "end_turn": "#/components/schemas/EndTurn",
                "maritime_trade": "#/components/schemas/MaritimeTrade",
                "move_robber": "#/components/schemas/MoveRobber",
                "play_knight": "#/components/schemas/PlayKnight",
                "play_monopoly": "#/components/schemas/PlayMonopoly",
                "play_road_building": "#/components/schemas/PlayRoadBuilding",
                "play_year_of_plenty": "#/components/schemas/PlayYearOfPlenty",
                "propose_trade": "#/components/schemas/ProposeTrade",
                "respond_trade": "#/components/schemas/RespondTrade",
                "roll_dice": "#/components/schemas/RollDice"
              }
            }
          }
        },
        "type": "object",
        "required": [
          "action"
        ],
        "title": "GameActionRequest",
        "description": "Request body for POST /games/{game_id}/actions.\n\nAttributes:\n    action: The move to make, discriminated by its `type`."
      },
      "GameCreate": {
        "properties": {
          "max_players": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 2.0,
            "title": "Max Players",
            "default": 4
          },
          "bots": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 0.0,
            "title": "Bots",
            "default": 0
          }
        },
        "type": "object",
        "title": "GameCreate",
        "description": "Request body for POST /games.\n\nAttributes:\n    max_players: Seats at the table, including the host and bots.\n    bots: Seats to fill with random bots straight away."
      },
      "GameDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/GameStatus"
          },
          "host_name": {
            "type": "string",
            "title": "Host Name"
          },
          "is_host": {
            "type": "boolean",
            "title": "Is Host"
          },
          "max_players": {
            "type": "integer",
            "title": "Max Players"
          },
          "seats": {
            "items": {
              "$ref": "#/components/schemas/GameSeat"
            },
            "type": "array",
            "title": "Seats"
          },
          "your_seat": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Your Seat"
          },
          "your_turn": {
            "type": "boolean",
            "title": "Your Turn"
          },
          "current_player_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Player Name"
          },
          "winner_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winner Name"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "view": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GameView"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "host_name",
          "is_host",
          "max_players",
          "seats",
          "your_seat",
          "your_turn",
          "current_player_name",
          "winner_name",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "GameDetail",
        "description": "A game with the requesting user's view of the board.\n\nAttributes:\n    view: What the requesting user may see; None while in the lobby."
      },
      "GameList": {
        "properties": {
          "mine": {
            "items": {
              "$ref": "#/components/schemas/GameSummary"
            },
            "type": "array",
            "title": "Mine"
          },
          "open": {
            "items": {
              "$ref": "#/components/schemas/GameSummary"
            },
            "type": "array",
            "title": "Open"
          }
        },
        "type": "object",
        "required": [
          "mine",
          "open"
        ],
        "title": "GameList",
        "description": "Response for GET /games.\n\nAttributes:\n    mine: Games the requesting user is seated in, most recently active first.\n    open: Lobbies with a free seat the requesting user could join."
      },
      "GameSeat": {
        "properties": {
          "seat": {
            "type": "integer",
            "title": "Seat"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "is_bot": {
            "type": "boolean",
            "title": "Is Bot"
          },
          "is_you": {
            "type": "boolean",
            "title": "Is You"
          }
        },
        "type": "object",
        "required": [
          "seat",
          "name",
          "is_bot",
          "is_you"
        ],
        "title": "GameSeat",
        "description": "One occupied seat, as shown in lobbies and game lists.\n\nAttributes:\n    seat: Seat number; the engine player id once the game starts.\n    name: Display name.\n    is_bot: Whether a random bot plays this seat.\n    is_you: Whether the requesting user holds this seat."
      },
      "GameStatus": {
        "type": "string",
        "enum": [
          "waiting",
          "active",
          "finished"
        ],
        "title": "GameStatus",
        "description": "Lifecycle of a game: lobby, in play, done."
      },
      "GameSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/GameStatus"
          },
          "host_name": {
            "type": "string",
            "title": "Host Name"
          },
          "is_host": {
            "type": "boolean",
            "title": "Is Host"
          },
          "max_players": {
            "type": "integer",
            "title": "Max Players"
          },
          "seats": {
            "items": {
              "$ref": "#/components/schemas/GameSeat"
            },
            "type": "array",
            "title": "Seats"
          },
          "your_seat": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Your Seat"
          },
          "your_turn": {
            "type": "boolean",
            "title": "Your Turn"
          },
          "current_player_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Player Name"
          },
          "winner_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Winner Name"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "host_name",
          "is_host",
          "max_players",
          "seats",
          "your_seat",
          "your_turn",
          "current_player_name",
          "winner_name",
          "version",
          "created_at",
          "updated_at"
        ],
        "title": "GameSummary",
        "description": "A game as listed in the lobby.\n\nAttributes:\n    id: Game id.\n    status: waiting, active, or finished.\n    host_name: Display name of the host.\n    is_host: Whether the requesting user is the host.\n    max_players: Seats at the table.\n    seats: Occupied seats in seat order.\n    your_seat: The requesting user's seat, if seated.\n    your_turn: Whether the game is waiting on the requesting user.\n    current_player_name: Whose turn it is, once started.\n    winner_name: Winner, once finished.\n    version: Bumped on every change.\n    created_at: When the game was created.\n    updated_at: When the game last changed."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MoveRobber": {
        "properties": {
          "type": {
            "type": "string",
            "const": "move_robber",
            "title": "Type",
            "default": "move_robber"
          },
          "hex_id": {
            "type": "integer",
            "title": "Hex Id"
          },
          "victim_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Victim Id"
          }
        },
        "type": "object",
        "required": [
          "hex_id"
        ],
        "title": "MoveRobber",
        "description": "move the robber and steal one card from a player next to its new hex."
      },
...
      "RollDice": {
        "properties": {
          "type": {
            "type": "string",
            "const": "roll_dice",
            "title": "Type",
            "default": "roll_dice"
          }
        },
        "type": "object",
        "title": "RollDice",
        "description": "roll both dice to start the main part of the turn."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}